requirements.txt
requests~=2.25.1
jieba~=0.42.1
matplotlib~=3.4.2
wordcloud~=1.8.1
Pillow~=8.2.0
PyPtt~=0.9.47